Break out common funcs
working with win/mac
